Removed all 'fracts', moved to Scalable interface

+++ src/io/anuke/ucore/entities/DestructibleEntity.java
@@ -3,9 +3,9 @@ package io.anuke.ucore.entities;
 import io.anuke.ucore.util.Mathf;
 
 public abstract class DestructibleEntity extends SolidEntity{
-	public transient int maxhealth;
+	public transient float maxhealth;
 	public transient boolean dead;
-	public int health;
+	public float health;
 	
 	public void onHit(SolidEntity entity){}
 	public void onDeath(){}
@@ -27,7 +27,7 @@ public abstract class DestructibleEntity extends SolidEntity{
 		}
 	}
 	
-	public void damage(int amount){
+	public void damage(float amount){
 		health -= amount;
 		if(health <= 0 && !dead){
 			onDeath();
@@ -35,7 +35,7 @@ public abstract class DestructibleEntity extends SolidEntity{
 		}
 	}
 	
-	public void setMaxHealth(int health){
+	public void setMaxHealth(float health){
 		maxhealth = health;
 		heal();
 	}
